requirements.txt
# feedparser is only if you want to scrape arxiv
feedparser
numpy
scipy

# sciki-learn is only for sparse arrays
scikit-learn

# twitter
tweepy
pytz

# the following are only for serve.py
python-dateutil
flask
flask_limiter
tornado
pymongo
